# myride/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MyRide",
        "description": "API Documentation for MyRide",
        "contact": {
            "email": "flazen.edu@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/question/faq": {
            "get": {
                "tags": [
                    "Question"
                ],
                "summary": "Get showing FAQ",
                "description": "This request is used to get showing FAQ in the welcome page. This request is using MySql database",
                "operationId": "27dc39035fffcbd37f6e60fbbf93d66f",
                "responses": {
                    "200": {
                        "description": "faq fetched",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "faq fetched"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "faq_question": {
                                                        "type": "string",
                                                        "example": "lorem ipsum?"
                                                    },
                                                    "faq_answer": {
                                                        "type": "string",
                                                        "example": "consectetuer adipiscing elit"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "faq failed to fetched",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "failed"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "faq not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "something wrong. please contact admin"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "JWT Authorization header using the Bearer scheme",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}